requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1;python_version<"3.9"
base58==2.1.1
billiard==3.6.4.0
bitarray==2.7.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.31.0
cytoolz==0.12.1
Django==4.1.7
django-celery-beat==2.4.0
django-crispy-forms==1.14.0
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
django-timezone-field==5.0
django-tinymce==3.5.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
exceptiongroup==1.1.0
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
importlib-resources==5.10.2
iniconfig==2.0.0
ipfshttpclient==0.8.0a2
Jinja2==3.1.2
Js2Py==0.74
jsonschema==4.17.3
kombu==5.2.4
lru-dict==1.1.8
MarkupSafe==2.1.2
multiaddr==0.0.9
multidict==6.0.4
netaddr==0.8.0
packaging==23.0
parsimonious==0.8.1
phonenumbers==8.13.5
pkgutil_resolve_name==1.3.10
pluggy==1.0.0
prompt-toolkit==3.0.36
protobuf==3.19.5
psycopg2-binary==2.9.5
pycryptodome==3.17
pyjsparser==2.7.1
pyrsistent==0.19.3
pytest==7.2.1
pytest-django==4.5.2
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
redis==4.4.2
requests==2.28.2
rlp==2.0.1
six==1.16.0
sqlparse==0.4.3
temp==2020.7.2
tomli==2.0.1
toolz==0.12.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
varint==1.0.2
vine==5.0.0
wcwidth==0.2.6
web3==5.31.3
websockets==9.1
yarl==1.8.2
zipp==3.12.1
